refactor(frontend): move atoms and molecules to its module
