--- packages/frontend/src/components/atoms/index.ts
@@ -1,6 +1,0 @@
-import LogOut from './LogOut';
-import UserInput from './UserInput';
-import UserSubmit from './UserSubmit';
-import MiniProduct from '../modules/User/MiniProduct';
-
-export { LogOut, UserInput, UserSubmit, MiniProduct };
