+++ packages/frontend/src/components/pages/index.ts
@@ -1,6 +1,3 @@
 import Main from './Main';
-import User from '../modules/User/User';
-import UserProducts from '../modules/User/userProductsPage/UserProducts';
-import UserCart from '../modules/User/userCarts/UserCart';
 
-export { Main, User, UserProducts, UserCart };
+export { Main };
